Overhaul tools and tutorials (#17)
# Overhaul of 3lc-examples tutorials

## Summary
This PR introduces a significant overhaul of the tutorials and src/tlc_tools folders to improve usability, structure, and clarity for users. The updates provide new tutorials, and a helper package which can be installed from source.

## Key Changes
New Tutorials Organized by Workflow
Ready-to-Use Toy Data
Introducing tlc_tools Library
Restructured Layout

## Motivation
The goal of this overhaul is to make the 3lc-examples repository a more accessible and user-friendly resource. It demonstrates 3LC capabilities across common workflows and lowers the barrier to entry for new users.
---------

Co-authored-by: Frederik Mellbye <frederik.mellbye@3lc.ai>
Co-authored-by: Markus Rognlien <markus.rognlien@3lc.ai>
Co-authored-by: Frederik J. Mellbye <20386018+fredmell@users.noreply.github.com>
Co-authored-by: Kelly Walker <kelly.walker@3lc.ai>