Use more descriptive test names for notebook tests (#38)
Previously, only the stem of the notebook path was used as a test name, which made it hard to determine which tests belong to the same group. Now, test names use notebook names relative to the root notebook folder.

Also add some preparation for adding tags to cell metadata, in order to filter out slow-running notebooks during testing. TBC.