Add .filename to PIL images written in add_columns_to_table tool (#41)
We need to convert values to "sample view" when adding input values to a derived table (using TableWriter).
This is unfortunate, but currently a limitation. When we use a BytesIO buffer to load the image data before opening it using Image.open, the filename information is lost, causing the TableWriter to attempt to write a copy of the image, and failing because of lack of bulk data prefix. This simple fix adds back the _tlc_url to the sample, allowing us to avoid file copying.